design: home으로 버튼 사이드바 탭으로 수정
