+++ today's horoscope/apps/admin/app/@types/index.d.ts
@@ -1,1 +1,1 @@
-type TabType = '관리자 설정' | '통계' | '기능 관리';
+type TabType = 'HOME' | '관리자 설정' | '통계' | '기능 관리';
