+++ today's horoscope/apps/admin/app/_components/header/Header.tsx
@@ -1,14 +1,14 @@
 import Image from 'next/image';
 import React from 'react';
 import { Logo } from '../../_images';
-import HeaderButtonSection from './HeaderButtonSection';
+import LogoutButton from './LogoutButton';
 
 function Header() {
   return (
     <header className="h-[150px] relative flex justify-center items-center">
       <Image src={Logo} alt="Logo" className="w-44" />
 
-      <HeaderButtonSection />
+      <LogoutButton />
     </header>
   );
 }
