+++ today's horoscope/apps/admin/app/_components/header/LogoutButton.tsx
@@ -1,0 +1,24 @@
+'use client';
+
+import React from 'react';
+import { usePathname, useRouter } from 'next/navigation';
+import routes from '../../_constants/routes';
+
+function LogoutButton() {
+  const pathname = usePathname();
+  const router = useRouter();
+
+  if (pathname === routes.LOGIN) return;
+
+  const handleLogout = () => {
+    router.push(routes.LOGIN);
+  };
+
+  return (
+    <button onClick={handleLogout} className="absolute bottom-5 right-20 flex gap-[60px] font-medium">
+      로그아웃
+    </button>
+  );
+}
+
+export default LogoutButton;
